% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_subsets.R
\name{all_subsets}
\alias{all_subsets}
\title{all_subsets Function}
\usage{
all_subsets(series, covariates, min, max, type, fit = TRUE)
}
\arguments{
\item{series}{Data with annual values of response (abundance) and predictor covariates.}

\item{covariates}{Vector of covariates to include in models.}

\item{min}{Minimum number of covariates within individual model.}

\item{max}{Maximum number of covariates within individual model.}

\item{type}{Specifies whether an in-season or pre-season forecast.}

\item{fit}{Flag for whether to fit models.}
}
\value{
A list containing ...
}
\description{
A comprehensive R package for time series analysis, forecasting, and regression modeling.
}
