% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dat.R
\name{brood_to_return}
\alias{brood_to_return}
\title{Convert a brood table to a return table}
\usage{
brood_to_return(bt)
}
\arguments{
\item{bt}{Brood table data frame.}
}
\value{
Returns a "return table" data frame.
}
\description{
Useful if you have data in brood table format because it needs to be in return table format for the "make_dat_function.
}
