% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning_checks.R
\name{check_nyrs}
\alias{check_nyrs}
\title{evaluate performance of using different numbers of years when calculating performance metrics for model selection and weighting}
\usage{
check_nyrs(
  One_ahead_out,
  dat,
  num_models,
  TY_ensemble,
  model_list,
  alpha,
  model_to_plot = "MAPE_weighted"
)
}
\arguments{
\item{dat}{data frame with columns "year", "species", "period", "abundance", and the covariates included in the covariates argument}

\item{num_models}{Number of top models to consider in ensemble creation.}

\item{TY_ensemble}{number of years ti evaluate performance based on plus 1.}

\item{alpha}{the annual rate of decay per year used in weighting performance in metric  calculation for model selection and ensemble-weights}
}
\value{
A list containing tables and plots summarizing the forecasting results and a list containing various outputs, including selected covariates, forecast results, rolling performance, ensemble models, and plots/tables.
}
\description{
Perform Forecasting with Ensemble Models
}
\details{
This function conducts time series forecasting using ensemble models.
}
