% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\docType{data}
\name{dat}
\alias{dat}
\title{Example dataset from the Yuan package}
\format{
A data frame with 100 rows and 12 columns.
}
\usage{
data(dat)
}
\description{
This dataset contains sample data for demonstration purposes.
}
\details{
The dataset includes the following variables:
\itemize{
\item \code{MEI.OND}: Description of MEI.OND variable.
\item \code{PDO.MJJ}: Description of PDO.MJJ variable.
\item \code{SSH.AMJ}: Description of SSH.AMJ variable.
\item \code{SST.AMJ}: Description of SST.AMJ variable.
\item \code{UWI.JAS}: Description of UWI.JAS variable.
\item \code{UWI.SON}: Description of UWI.SON variable.
\item \code{WSST_A}: Description of WSST_A variable.
\item \code{abundance}: Description of abundance variable.
\item \code{lag1_NPGO}: Description of lag1_NPGO variable.
\item \code{lag1_PDO}: Description of lag1_PDO variable.
\item \code{lag1_log_JackOPI}: Description of lag1_log_JackOPI variable.
\item \code{lag1_log_SmAdj}: Description of lag1_log_SmAdj variable.
\item \code{period}: Description of period variable.
\item \code{species}: Description of species variable.
\item \code{year}: Description of year variable.
}
}
\examples{
data(dat)
}
\keyword{datasets}
