% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_forecast.R
\name{do_forecast}
\alias{do_forecast}
\title{Perform Forecasting with Ensemble Models}
\usage{
do_forecast(
  covariates = c("lag1_log_JackOPI", "lag1_log_SmAdj", "lag1_NPGO", "lag1_PDO", "WSST_A",
    "PDO.MJJ", "MEI.OND", "UWI.JAS", "SST.AMJ", "SSH.AMJ", "UWI.SON"),
  leave_yrs = 31,
  TY_ensemble = 16,
  first_forecast_period = 1,
  plot_results = FALSE,
  write_model_summaries = TRUE,
  forecast_period_start_m = 1,
  forecast_period_start_d = 1,
  stack_metric = "MAPE",
  k = 1,
  min_vars = 0,
  max_vars = 1,
  forecast_type = "preseason",
  rolling_year_window = 15,
  num_models = 10
)
}
\arguments{
\item{covariates}{A vector specifying the covariates to be considered in the forecasting models.}

\item{leave_yrs}{Number of years to leave out for model validation.}

\item{TY_ensemble}{Number of models in the ensemble.}

\item{first_forecast_period}{The starting period for making forecasts.}

\item{plot_results}{Logical, indicating whether to plot forecast results.}

\item{write_model_summaries}{Logical, indicating whether to write model summaries.}

\item{forecast_period_start_m}{Starting month of the forecast period (inclusive).}

\item{forecast_period_start_d}{Starting day of the forecast period (inclusive).}

\item{stack_metric}{The metric used for stacking models in the ensemble.}

\item{k}{Number of top models to include in the ensemble.}

\item{min_vars}{Minimum number of covariates in a model.}

\item{max_vars}{Maximum number of covariates in a model.}

\item{forecast_type}{Type of forecasting approach ("preseason" or other).}

\item{rolling_year_window}{Number of years for rolling performance calculation.}

\item{num_models}{Number of top models to consider in ensemble creation.}
}
\value{
A list containing tables and plots summarizing the forecasting results and a list containing various outputs, including selected covariates, forecast results, rolling performance, ensemble models, and plots/tables.
}
\description{
Perform Forecasting with Ensemble Models
}
\details{
This function conducts time series forecasting using ensemble models.
}
