% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{ensemble}
\alias{ensemble}
\title{Function to perform ensemble analysis}
\usage{
ensemble(forecasts, series, TY_ensemble, k, slide, num_models, stack_metric)
}
\arguments{
\item{forecasts}{A data frame containing forecasts data.}

\item{series}{A data frame containing series data.}

\item{TY_ensemble}{An integer indicating the number of years for the ensemble analysis.}

\item{k}{A numeric value for the weighting parameter.}

\item{slide}{An integer indicating the length of the forecast slide.}
}
\value{
A list of results, including final_model_weights, forecast_skill, ensembles, and forecasts.
}
\description{
This function takes forecasts, series, TY_ensemble, k, and slide as inputs and performs ensemble analysis.
}
