% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{evaluate_forecasts2}
\alias{evaluate_forecasts2}
\title{Calculate performance metrics of models}
\usage{
evaluate_forecasts2(forecasts)
}
\arguments{
\item{forecasts}{dataframe of one-step-ahead forecasts}
}
\value{
a data frame with  performance metrics calculated for each model
}
\description{
Calculate performance metrics of models
}
