% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_stack_weights.R
\name{find_stack_weights}
\alias{find_stack_weights}
\title{Optimization algorithm to find best stacking weights}
\usage{
find_stack_weights(tau, metric, n, initial_weights, preds, obs)
}
\arguments{
\item{tau}{Tuning parameter.}

\item{metric}{Metric for evaluation.}

\item{n}{Number of iterations.}

\item{initial_weights}{Initial weights.}

\item{preds}{Matrix of predictions.}

\item{obs}{Vector of observations.}
}
\value{
List containing weights, skill, and skill_list.
}
\description{
Optimization algorithm to find best stacking weights
}
