% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dat_from_excel.R
\name{make_dat_from_excel}
\alias{make_dat_from_excel}
\title{Make dat from Excel}
\usage{
make_dat_from_excel(excel_path, file_path = NULL, redo = TRUE)
}
\arguments{
\item{excel_path}{Path to the Excel file.}

\item{file_path}{Path to an existing file or NULL.}

\item{redo}{Boolean indicating whether to remake the file.}
}
\value{
A data frame with the processed data.
}
\description{
Convert a brood table to a return table and scrape covariates from the web.
}
