% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_step_ahead.R
\name{one_step_ahead}
\alias{one_step_ahead}
\title{Evaluate ARIMA model performance}
\usage{
one_step_ahead(
  series,
  leave_yrs,
  TY_ensemble,
  covariates,
  first_forecast_period
)
}
\arguments{
\item{series}{The input time series data.}

\item{leave_yrs}{The number of years to leave out for testing.}

\item{TY_ensemble}{The ensemble method to use}

\item{covariates}{Explanatory variables matrix for model fitting.}

\item{first_forecast_period}{The first forecast period.}
}
\value{
A list with forecast results including 'pred', 'CI', 'arma', and 'aicc'.
}
\description{
This function evaluates the performance of an ARIMA model for seasonal total forecasts.
}
\examples{
\dontrun{
  result <- one_step_ahead(...)
}

}
