% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{rolling_perf}
\alias{rolling_perf}
\title{Function to calculate performance on a rolling window}
\usage{
rolling_perf(
  one_aheads,
  series,
  roll_years = 15,
  mod_include = 5,
  TY_ensemble,
  forecast_year
)
}
\arguments{
\item{one_aheads}{a data frame containing forecasts}

\item{series}{a data frame containing observations}

\item{roll_years}{number of years within the rolling window}

\item{mod_include}{number of top models to return in an output table}

\item{TY_ensemble}{number of years of performance to look at when selecting top models to return in output table}

\item{forecast_year}{the final year of forecasts, for which an observationis not available}
}
\value{
a list with three components
}
\description{
This function calculates the performance of a series of forecasts over a rolling (i.e., sliding) window.
}
