% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima_forecast.R
\name{arima_forecast}
\alias{arima_forecast}
\title{ARIMA Forecast Function}
\usage{
arima_forecast(
  tdat,
  xreg,
  xreg_pred,
  last_train_yr,
  first_forecast_period,
  write_model_summaries,
  train_test,
  abundance
)
}
\arguments{
\item{tdat}{The time series data.}

\item{xreg}{The exogenous variables used in the model.}

\item{xreg_pred}{The exogenous variables for prediction.}

\item{last_train_yr}{The last year of training data.}

\item{first_forecast_period}{The first forecast period.}

\item{write_model_summaries}{A logical indicating whether to write model summaries.}

\item{train_test}{A vector or column of the time series data indicating training/test periods.}

\item{abundance}{A vector or column of the time series data indicating abundance values.}
}
\value{
A description of the return value.
}
\description{
A comprehensive R package for time series analysis, forecasting, and regression modeling.
}
