% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{ensemble}
\alias{ensemble}
\title{Ensemble Forecasting Function}
\usage{
ensemble(
  forecasts,
  series,
  TY_ensemble,
  k,
  slide,
  num_models,
  stack_metric,
  stretch = FALSE
)
}
\arguments{
\item{forecasts}{A data frame containing model forecasts.}

\item{series}{A data frame containing observed time series data.}

\item{TY_ensemble}{The number of years to consider for ensemble generation.}

\item{k}{A parameter for weighting in the ensemble generation.}

\item{slide}{The length of the sliding window for calculating ensemble weights}

\item{num_models}{The number of top models to consider in each iteration.}

\item{stack_metric}{The metric to use for stacking weights.}

\item{stretch}{Boolean whether to stretch rather than slide}
}
\value{
A list containing final model weights, forecast skill evaluation, ensembles, and updated forecasts.
}
\description{
A function that generates ensemble forecasts based on input forecasts and series data.
}
