% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_stack_weights.R
\name{find_stack_weights}
\alias{find_stack_weights}
\title{Find Best Stacking Weights}
\usage{
find_stack_weights(tau, metric, n, initial_weights, preds, obs)
}
\arguments{
\item{tau}{Parameter for the Dirichlet distribution}

\item{metric}{Performance metric to optimize (e.g., "RMSE", "MAPE")}

\item{n}{Number of iterations}

\item{initial_weights}{Initial weights for stacking}

\item{preds}{Matrix of predictions from base models}

\item{obs}{Vector of observed values}
}
\value{
A list containing optimized weights, best skill metric, and skill metric history
}
\description{
This function implements an optimization algorithm to find the best stacking weights for a set of base models, given observed values.
}
