% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dat.R
\name{make_dat}
\alias{make_dat}
\title{Make data}
\usage{
make_dat(
  dat1,
  covariates = c("lag1_log_Jack", "lag4_log_adults", "lag5_log_adults", "lag1_log_SAR",
    "lag2_log_SAR", "lag1_NPGO", "lag1_PDO", "lag2_NPGO", "lag2_PDO", "lag2_PC1",
    "lag2_PC2", "lag2_sp_phys_trans", "pink_ind", "lag1_log_socksmolt")
)
}
\arguments{
\item{dat1}{Input file with adult and jack returns by year.}

\item{covariates}{Vector of covariates to include in the output.}
}
\value{
A data frame with the processed data.
}
\description{
Scrape covariates from the web and add covariates to return table
}
