% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_step_ahead.R
\name{one_step_ahead}
\alias{one_step_ahead}
\title{Function to evaluate the performance of ARIMA model (produces season total forecasts only)}
\usage{
one_step_ahead(
  series,
  leave_yrs,
  TY_ensemble,
  covariates,
  first_forecast_period,
  ts_freq = 1,
  seasonal = FALSE,
  plot_results,
  write_model_summaries,
  forecast_period_start_m,
  forecast_period_start_d,
  stack_metric,
  k
)
}
\arguments{
\item{series}{A data frame containing the time series data.}

\item{leave_yrs}{Number of years to leave out during the evaluation.}

\item{TY_ensemble}{A parameter for ensemble evaluation.}

\item{covariates}{A list of covariates to include in the model.}

\item{first_forecast_period}{The first period for forecasting.}

\item{ts_freq}{frequency of time series being forecasts}

\item{seasonal}{whether the arima models should include seasonal  components}

\item{plot_results}{A logical value to indicate whether to plot results.}

\item{write_model_summaries}{A logical value to indicate whether to write model summaries.}

\item{forecast_period_start_m}{The starting month for the forecast period (inclusive).}

\item{forecast_period_start_d}{The starting day for the forecast period (inclusive).}

\item{stack_metric}{A metric used for stacking models.}

\item{k}{A parameter for the model evaluation.}

\item{alpha}{the annual rate of decay per year used in weighting performance in metric  calculation for weight calculation for ensemble}
}
\value{
A data frame containing forecasts and other information.
}
\description{
This function evaluates the performance of an ARIMA model for season total forecasts using provided covariates and other parameters.
}
