% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_table.R
\name{plot_table}
\alias{plot_table}
\title{Plot and Generate Tables for Forecasting Results
This function creates tables and plots to summarize the forecasting results.}
\usage{
plot_table(rp, ens, stack_metric, rolling_year_window, output_path = "outputs")
}
\arguments{
\item{rp}{Rolling performance data.}

\item{ens}{Ensemble model data.}
}
\value{
A list containing tables and plots summarizing the forecasting results.
}
\description{
Plot and Generate Tables for Forecasting Results
}
