% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_table.R
\name{plot_table}
\alias{plot_table}
\title{Plot and Generate Tables for Forecasting Results
This function creates tables and plots to summarize the forecasting results.}
\usage{
plot_table(
  rp,
  ens,
  dat,
  stack_metric,
  rolling_year_window,
  benchmark = NULL,
  output_path = "outputs"
)
}
\arguments{
\item{rp}{Rolling performance data.}

\item{ens}{Ensemble model data.}

\item{dat}{the input data with columns "year" and "abundance"}

\item{benchmark}{info on a  model to be compared to}
}
\value{
A list containing tables and plots summarizing the forecasting results.
}
\description{
Plot and Generate Tables for Forecasting Results
}
