% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_perf.R
\name{rolling_perf}
\alias{rolling_perf}
\title{Calculate rolling performance of models}
\usage{
rolling_perf(
  one_aheads,
  series,
  roll_years,
  mod_include,
  TY_ensemble,
  model_list
)
}
\arguments{
\item{one_aheads}{A data frame with one-step ahead model predictions.}

\item{series}{A data frame with time series data that includes 'year' and 'abundance' columns.}

\item{roll_years}{The number of years for rolling calculation (default is 15).}

\item{mod_include}{The number of top-performing models to include (default is 5).}

\item{TY_ensemble}{The number of years for the ensemble analysis (default: 16).}

\item{model_list}{A data frame with information about each model, including the 'model' column.}
}
\value{
A list with three elements:
\itemize{
\item 'all_mods': A data frame with all models' performance metrics.
\item 'top_mods': A data frame with the top-performing models' performance metrics.
\item 'performance': A summary of the performance metrics for the top models.
}
}
\description{
This function calculates the rolling performance of models on time series data.

This function calculates the rolling performance of models on time series data.
}
